+++ app/src/main/kotlin/paufregi/connectfeed/data/database/entities/GearEntity.kt
@@ -12,5 +12,5 @@ data class GearEntity(
     val userId: Long,
     val name: String,
     val type: GearType = GearType.Unknown,
-    val distance: Long? = null,
+    val distance: Int? = null,
 )
